+++ LoanDecider/Loan.cs
@@ -2,18 +2,18 @@
 {
     public class Loan
     {
-        private decimal rate;
+        private double rate;
         private decimal monthlyRepayment;
         private decimal totalRepayment;
 
-        public Loan(decimal rate, decimal monthlyRepayment, decimal totalRepayment)
+        public Loan(double rate, decimal monthlyRepayment, decimal totalRepayment)
         {
             this.rate = rate;
             this.monthlyRepayment = monthlyRepayment;
             this.totalRepayment = totalRepayment;
         }
 
-        public decimal Rate { get => rate; }
+        public double Rate { get => rate; }
         public decimal MonthlyRepayment { get => monthlyRepayment;  }
         public decimal TotalRepayment { get => totalRepayment;  }
     }
